MA-diffuison/requirements.txt
appdirs==1.4.4
attrs==19.3.0
Automat==0.8.0
black==19.3b0
blinker==1.4
blobfile==1.2.9
certifi==2019.11.28
cftime==1.6.2
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
einops==0.6.0
entrypoints==0.3
filelock==3.0.12
flake8==5.0.4
fonttools==4.34.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
isort==5.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
MarkupSafe==1.1.0
matplotlib==3.5.2
mccabe==0.7.0
more-itertools==4.2.0
netCDF4==1.6.2
netifaces==0.10.4
netron==6.7.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
packaging==21.3
pandas==1.5.2
parameterized==0.8.1
Pillow==9.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.9.1
pycryptodomex==3.17
pyflakes==2.5.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrealsense2==2.50.0.3812
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.9.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
Twisted==18.9.0
typing_extensions==4.3.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
vit-pytorch==1.2.0
wadllib==1.3.3
wget==3.2
xmltodict==0.12.0
yacs==0.1.8
zipp==1.0.0
zope.interface==4.7.1
